Adiciona tratamento de exceção para método delete do service

+++ src/main/java/br/com/devsuperior/clientmanager/services/ClientService.java
@@ -54,6 +54,9 @@ public class ClientService {
 
   @Transactional
   public void delete(Long id) {
+    if(!repository.existsById(id)){
+      throw new ResourceNotFoundException("Recurso não encontrado");
+    }
     repository.deleteById(id);
   }
